fix(journal): split filesystem journal loading out of the parser
The raw-text entry lens is a client component that imports parseBlock from
lib/journal/parser. That module also held resolveIncludes and parseJournal,
which import 'fs'/'path', so bundling the client pulled Node built-ins into
the browser graph and the Turbopack build failed with "Can't resolve 'fs'".

Move the two filesystem-dependent functions into a new server-only
lib/journal/loader module, leaving parser pure and browser-safe. Repository
and service now import the loader; the include/journal tests move alongside
into loader.test. Pure parsing keeps its existing home and tests.
