+++ lib/journal/repository.ts
@@ -3,7 +3,8 @@ import { promises as fs } from 'fs';
 import path from 'path';
 import { eq, sql } from 'drizzle-orm';
 import { DEFAULT_MAIN, PRICE_DB_NAME, getJournalDir } from './layout';
-import { parseJournal, type ParsedJournal, type Transaction } from './parser';
+import { parseJournal } from './loader';
+import { type ParsedJournal, type Transaction } from './parser';
 import { userSetting } from '@/db/schema';
 import type { DbInstance } from '@/lib/db/connection';
 import { pullLocked, manifestRelName } from '@/lib/storage';
